Revert "Channel overrides (#53)"
This reverts commit d6c1f5cd43d69890b05a261c232107c7681dbec5.
