--- src/Discord/Commands/stats.js
@@ -1,12 +1,0 @@
-require("dotenv").config()
-import admin from "firebase-admin";
-
-module.exports = {
-	name: "stats",
-	aliases: [],
-    description: "Get DisStreamChat Stats",
-	adminOnly: true,
-	execute: async (message, args, client) => {
-		// const users = /
-	},
-};
