+++ src/Discord/Events/channelDelete.js
@@ -1,12 +1,17 @@
 import admin from "firebase-admin";
 import { MessageEmbed } from "discord.js";
-import setupLogging from "./utils/setupLogging";
 
 module.exports = async (channel, client) => {
 	const guild = channel.guild;
 
-    const [channelId, active] = await setupLogging(guild, "channelDelete")
-    if(!active) return
+	let channelId = null;
+	const serverRef = await admin.firestore().collection("loggingChannel").doc(guild.id).get();
+	const serverData = serverRef.data();
+	if (serverData) {
+        channelId = serverData.server;
+        const activeLogging = serverData.activeEvents || {}
+        if(!activeLogging["channelDelete"]) return 
+	}
 
 	let parentCheck = '';
 	if (channel.parentID) {
