+++ src/Discord/Events/emojiCreate.js
@@ -1,16 +1,18 @@
 import admin from "firebase-admin";
 import { MessageEmbed } from "discord.js";
-import setupLogging from "./utils/setupLogging";
 
 module.exports = async emoji => {
     const guild = emoji.guild;
     if(!guild) return
 
-
-    const [channelId, active] = await setupLogging(guild, "emojiCreate")
-    if(!active) return
-
-
+	let channelId = null;
+	const serverRef = await admin.firestore().collection("loggingChannel").doc(guild.id).get();
+	const serverData = serverRef.data();
+	if (serverData) {
+        channelId = serverData.server;
+        const activeLogging = serverData.activeEvents || {}
+        if(!activeLogging["emojiCreate"]) return 
+    }
     
     const createdBy = await emoji.fetchAuthor()
