+++ src/Discord/Events/guildMemberAdd.js
@@ -1,14 +1,19 @@
 import admin from "firebase-admin";
 const { formatFromNow } = require("../../utils/functions");
 import { MessageEmbed } from "discord.js";
-import setupLogging from "./utils/setupLogging";
 
 module.exports = async (member, client) => {
 	const guild = member.guild;
 
-    const [channelId, active] = await setupLogging(guild, "MemberAdd")
-    if(!active) return
-    
+	let channelId = null;
+	const serverRef = await admin.firestore().collection("loggingChannel").doc(guild.id).get();
+	const serverData = serverRef.data();
+	if (serverData) {
+        channelId = serverData.server;
+        const activeLogging = serverData.activeEvents || {}
+        if(!activeLogging["MemberAdd"]) return 
+	}
+
 	const embed = new MessageEmbed()
 		.setAuthor(member.user.tag, member.user.displayAvatarURL())
 		.setThumbnail(member.user.displayAvatarURL())
