+++ src/Discord/Events/inviteDelete.js
@@ -1,13 +1,18 @@
 import admin from "firebase-admin";
 import { MessageEmbed } from "discord.js";
-import setupLogging from "./utils/setupLogging";
 
 module.exports = async invite => {
     const guild = invite.guild;
     if(!guild) return
 
-    const [channelId, active] = await setupLogging(guild, "InviteDelete")
-    if(!active) return
+	let channelId = null;
+	const serverRef = await admin.firestore().collection("loggingChannel").doc(guild.id).get();
+	const serverData = serverRef.data();
+	if (serverData) {
+        channelId = serverData.server;
+        const activeLogging = serverData.activeEvents || {}
+        if(!activeLogging["InviteDelete"]) return 
+	}
 
 	const embed = new MessageEmbed()
 		.setAuthor("DisStreamBot")
